conda-requirements.txt
tornado
pygithub
nose
mock
git
